+++ R/bam_to.R
@@ -1,9 +1,11 @@
-#' BAM to bedGraph
+#' Convert a BAM file
 #' 
 #' Convert a BAM file to any format supported by \link[rtracklayer]{export}.
+#' @param ... Additional arguments passed to \link[rtracklayer]{export}.
 #' @inheritParams pooled_peaks
 #' @inheritParams GenomicRanges::resize 
 #' @returns Converted file paths.
+#' 
 #' @export
 #' @importFrom stats setNames
 #' @importFrom rtracklayer import export
@@ -11,12 +13,13 @@
 #' @examples
 #' bam_files <- c(f1=system.file("extdata","querybins.bam",package="Rsamtools"),
 #'                f2=system.file("extdata","revbins.bam",package="Rsamtools"))
-#' converted_files <- convert_bam(bam_files = bam_files)               
-convert_bam <- function(bam_files,
-                        formats=c("bedGraph","bigWig"),
-                        width=150,
-                        outdir=NULL,
-                        verbose=TRUE){
+#' converted_files <- bam_to(bam_files = bam_files)               
+bam_to <- function(bam_files,
+                   formats=c("bedGraph","bigWig"),
+                   width=150,
+                   outdir=NULL,
+                   verbose=TRUE,
+                   ...){
     
     #### Name files ####
     if(is.null(names(bam_files))){
@@ -43,16 +46,16 @@ convert_bam <- function(bam_files,
         #### export rpm to... ####
         lapply(stats::setNames(formats,
                                formats), 
-               function(format){
-            outfile <- paste(gsub("\\.bam", "", file),format,sep=".")
+               function(f){
+            outfile <- paste(gsub("\\.bam", "", file),f,sep=".")
             if(!is.null(outdir)) {
                 outfile <- file.path(outdir,basename(outfile))
             }
-            messager(paste0("Exporting to ",format,":"),outfile,v=verbose)
+            messager(paste0("Exporting to ",f,":"),outfile,v=verbose)
             rtracklayer::export(object = rpm, 
                                 con = outfile, 
-                                format = format,
-                                ignore.strand = TRUE)
+                                format = f,
+                                ...)
             return(outfile)
         })  
     })